Settings view implemented
Settings view is now synchronised with a properties file. Settings view stores user choices for music during play, tetriminos dropdown speed, and tetriminos colors.

+++ src/fr/xnxa/tetrix/Score.java
@@ -1,10 +1,10 @@
 package fr.xnxa.tetrix;
 
 public class Score implements Comparable<Score> {
-	
+
 	private String playername;
 	private long points;
-	
+
 	public Score(String playername, long points) {
 		this.playername = playername.trim();
 		this.points = points;
@@ -20,9 +20,9 @@ public class Score implements Comparable<Score> {
 
 	@Override
 	public int compareTo(Score other) {
-		return ((Long) this.points).compareTo((Long) other.points);
+		return ((Long) this.points).compareTo(other.points);
 	}
-	
+
 	@Override
 	public String toString() {
 		return playername + ":" + points ;
